requirements/requirements.txt
tqdm
timm
lpips
torch
monai
pandas
pyyaml
pillow
easydict
objprint
diffusers
matplotlib 
scikit-image
transformers
opencv-python
huggingface_hub==0.36.0
segment_anything
accelerate==1.12.0
